weibo.cn/python/requirements.txt
tensorflow >= 1.0
opencv-python
scikit-learn
scipy
numpy
matplotlib
Pillow
requests
bs4